feat: complete float16_allowed integration and documentation
Added float16_allowed parameter to all wrapper functions and updated
comprehensive documentation for encoding system.

## Python Changes

### save_gsplats() Function (gsplats/io/save_gsplats.py:71)
- Added float16_allowed: bool = False parameter
- Passes to ArrayEncoder for dtype selection
- Enables explicit control of float16 vs float32 encoding

### Test Updates (3 files)
- test_save_load.py: Expects float32 in MEMORY mode (new default)
- test_format.py: Expects float32 for semantic type test
- test_dtype_support.py: Expects float32 for dtype config test

## Documentation Added

### encoding/README.md
- New "Compatibility Control" section
- Documents float16_allowed parameter usage
- Examples for TypeScript compatibility vs Python-only workflows
- Lists affected semantic types

### data/SPECIFICATIONS.md
- New "Actual Python Encoding Names" section
- Documents specific names: lut_uint8, rgb_uint8, broadcasted, etc.
- Explains TypeScript decoder matching strategy
- Clarifies implicit bounds for RGB/HDR quantization

## Test Results

Python: ✅ 1320 PASS, 0 FAIL, 6 SKIP
TypeScript: ✅ 12 PASS, 0 FAIL, 3 SKIP

All encoding tests verify compatibility between Python encoder and
TypeScript decoder with new float16_allowed architecture.

+++ packages/luxar/src/luxar/gsplats/io/tests/test_format.py
@@ -377,9 +377,9 @@ class TestFormatCompliance:
             splats_group = root["splats"]
 
             # Check semantic types via encoding names (spec Section "Semantic Types")
-            # Centers: COORDINATE → float16 in MEMORY mode
+            # Centers: COORDINATE → float32 in MEMORY mode (new default with float16_allowed=False)
             centers_enc = splats_group["centers"].attrs.get("encoding", {})
-            assert centers_enc["name"] == "float16"
+            assert centers_enc["name"] == "float32"
 
             # Sharpnesses: BOUNDED_SCALAR → bounded_scalar_uint8 in MEMORY mode
             sharpness_enc = splats_group["sharpnesses"].attrs.get("encoding", {})
